refactor(AbstractJdbcTemplate): 불필요한 구현 분리
