+++ app/src/main/java/com/techcourse/dao/JdbcUserDao.java
@@ -1,14 +1,29 @@
 package com.techcourse.dao;
 
-import com.interface21.jdbc.core.JdbcTemplate;
+import com.interface21.jdbc.core.AbstractJdbcTemplate;
+import com.interface21.jdbc.core.PrepareStatementSetter;
+import com.interface21.jdbc.core.RowMapper;
 import com.techcourse.domain.User;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 import javax.sql.DataSource;
 
 public class JdbcUserDao {
+
+    private static final RowMapper<User> USER_ROW_MAPPER = resultSet -> new User(
+            resultSet.getLong("id"),
+            resultSet.getString("account"),
+            resultSet.getString("password"),
+            resultSet.getString("email")
+    );
+
+    private static final PrepareStatementSetter STATEMENT_SETTER = (preparedStatement, objects) -> {
+        if (objects == null) {
+            return;
+        }
+        for (int index = 0; index < objects.length; index++) {
+            preparedStatement.setObject(index + 1, objects[index]);
+        }
+    };
 
     private final DataSource dataSource;
 
@@ -16,23 +31,8 @@ public class JdbcUserDao {
         this.dataSource = dataSource;
     }
 
-    public JdbcUserDao(final JdbcTemplate jdbcTemplate) {
-        this.dataSource = null;
-    }
-
     public void insert(final User user) {
         AbstractJdbcTemplate jdbcTemplate = new AbstractJdbcTemplate(dataSource) {
-            @Override
-            protected void setValues(PreparedStatement preparedStatement) throws SQLException {
-                preparedStatement.setString(1, user.getAccount());
-                preparedStatement.setString(2, user.getPassword());
-                preparedStatement.setString(3, user.getEmail());
-            }
-
-            @Override
-            protected Object mapRow(ResultSet resultSet) throws SQLException {
-                return null;
-            }
 
             @Override
             protected String createQuery() {
@@ -44,23 +44,11 @@ public class JdbcUserDao {
                 return dataSource;
             }
         };
-        jdbcTemplate.update();
+        jdbcTemplate.update(STATEMENT_SETTER, user.getAccount(), user.getPassword(), user.getEmail());
     }
 
     public void update(final User user) {
         AbstractJdbcTemplate jdbcTemplate = new AbstractJdbcTemplate(dataSource) {
-            @Override
-            protected void setValues(PreparedStatement preparedStatement) throws SQLException {
-                preparedStatement.setString(1, user.getAccount());
-                preparedStatement.setString(2, user.getPassword());
-                preparedStatement.setString(3, user.getEmail());
-                preparedStatement.setLong(4, user.getId());
-            }
-
-            @Override
-            protected Object mapRow(ResultSet resultSet) throws SQLException {
-                return null;
-            }
 
             @Override
             protected String createQuery() {
@@ -72,99 +60,52 @@ public class JdbcUserDao {
                 return dataSource;
             }
         };
-        jdbcTemplate.update();
+        jdbcTemplate.update(STATEMENT_SETTER, user.getAccount(), user.getPassword(), user.getEmail(), user.getId());
     }
 
     public List<User> findAll() {
         AbstractJdbcTemplate jbcTemplate = new AbstractJdbcTemplate(dataSource) {
-            @Override
-            protected void setValues(PreparedStatement preparedStatement) throws SQLException {
-                /*NOOP*/
-            }
-
             @Override
             protected String createQuery() {
                 return "SELECT id, account, password, email FROM users";
             }
 
             @Override
-            protected Object mapRow(ResultSet resultSet) throws SQLException {
-                return new User(
-                        resultSet.getLong("id"),
-                        resultSet.getString("account"),
-                        resultSet.getString("password"),
-                        resultSet.getString("email")
-                );
-            }
-
-            @Override
             protected DataSource getDataSource() {
                 return dataSource;
             }
         };
-        return jbcTemplate.queryList().stream()
-                .map(o -> (User) o)
-                .toList();
+        return jbcTemplate.queryList(USER_ROW_MAPPER, STATEMENT_SETTER);
     }
 
     public User findById(final Long id) {
         AbstractJdbcTemplate jbcTemplate = new AbstractJdbcTemplate(dataSource) {
-            @Override
-            protected void setValues(PreparedStatement preparedStatement) throws SQLException {
-                preparedStatement.setLong(1, id);
-            }
-
             @Override
             protected String createQuery() {
                 return "SELECT id, account, password, email FROM users WHERE id = ?";
             }
 
             @Override
-            protected Object mapRow(ResultSet resultSet) throws SQLException {
-                return new User(
-                        resultSet.getLong("id"),
-                        resultSet.getString("account"),
-                        resultSet.getString("password"),
-                        resultSet.getString("email")
-                );
-            }
-
-            @Override
             protected DataSource getDataSource() {
                 return dataSource;
             }
         };
-        return (User) jbcTemplate.query();
+        return jbcTemplate.query(USER_ROW_MAPPER, STATEMENT_SETTER, id);
     }
 
     public User findByAccount(final String account) {
         AbstractJdbcTemplate jbcTemplate = new AbstractJdbcTemplate(dataSource) {
-            @Override
-            protected void setValues(PreparedStatement preparedStatement) throws SQLException {
-                preparedStatement.setString(1, account);
-            }
-
             @Override
             protected String createQuery() {
                 return "SELECT id, account, password, email FROM users WHERE account=?";
             }
 
             @Override
-            protected Object mapRow(ResultSet resultSet) throws SQLException {
-                return new User(
-                        resultSet.getLong(1),
-                        resultSet.getString(2),
-                        resultSet.getString(3),
-                        resultSet.getString(4)
-                );
-            }
-
-            @Override
             protected DataSource getDataSource() {
                 return dataSource;
             }
         };
-        return (User) jbcTemplate.query();
+        return jbcTemplate.query(USER_ROW_MAPPER, STATEMENT_SETTER, account);
     }
 
     public DataSource getDataSource() {
